+++ libraryprojectjava1/src/main/java/com/example/libraryprojectjava1/service/TransactionService.java
@@ -2,11 +2,15 @@ package com.example.libraryprojectjava1.service;
 
 import com.example.libraryprojectjava1.pojo.entity.Transaction;
 
-import java.util.*;
+import java.util.List;
+import java.util.Optional;
 
 public interface TransactionService {
-    List<Transaction> getTransactionsByMember(Long memberId);
-    List<Transaction> getTransactionsByBook(Long bookId);
+    List<Transaction> getTransactionsByMember(Integer memberId);
+
+    List<Transaction> getTransactionsByBook(Integer bookId);
+
     Transaction saveTransaction(Transaction transaction);
-    Optional<Transaction> getTransactionById(Long id);
+
+    Optional<Transaction> getTransactionById(Integer id);
 }
